+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/global/model/UiReviewData.kt
@@ -1,4 +1,4 @@
-package com.avengers.nibobnebob.presentation.ui.global.model
+package com.avengers.nibobnebob.presentation.ui.main.global.model
 
 data class UiReviewData(
     val reviewId : Int,
