+++ Aos/app/src/main/res/layout/fragment_add_my_restaurant.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.avengers.nibobnebob.presentation.ui.global.restaurantadd.AddMyRestaurantViewModel" />
+            type="com.avengers.nibobnebob.presentation.ui.main.global.restaurantadd.AddMyRestaurantViewModel" />
 
     </data>
