+++ Aos/app/src/main/res/layout/fragment_restaurant_detail.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.avengers.nibobnebob.presentation.ui.global.restaurantdetail.RestaurantDetailViewModel" />
+            type="com.avengers.nibobnebob.presentation.ui.main.global.restaurantdetail.RestaurantDetailViewModel" />
 
     </data>
