+++ Aos/app/src/main/res/navigation/nav_graph.xml
@@ -104,7 +104,7 @@
     </fragment>
     <fragment
         android:id="@+id/addMyRestaurantFragment"
-        android:name="com.avengers.nibobnebob.presentation.ui.global.restaurantadd.AddMyRestaurantFragment"
+        android:name="com.avengers.nibobnebob.presentation.ui.main.global.restaurantadd.AddMyRestaurantFragment"
         android:label="AddMyRestaurantFragment" >
         <argument
             android:name="restaurantName"
@@ -118,7 +118,7 @@
 
     <fragment
         android:id="@+id/restaurantDetailFragment"
-        android:name="com.avengers.nibobnebob.presentation.ui.global.restaurantdetail.RestaurantDetailFragment"
+        android:name="com.avengers.nibobnebob.presentation.ui.main.global.restaurantdetail.RestaurantDetailFragment"
         android:label="RestaurantDetailFragment">
         <argument
             android:name="restaurantId"
